refactor shared
